работа с голосовыми сообщениями и mp3

+++ bot/main.py
@@ -1,12 +1,16 @@
 import os
 import telebot
 from telebot import types
+from pydub import AudioSegment
 
 bot = telebot.TeleBot("6189522732:AAE6lLTknaQcz008aDRykaiFrkwnE1OJAN8", parse_mode=None)
 split_string = 'SL96illlSfCB5zP'
 files = []
 if not os.path.isdir('recieved_files'):
     os.mkdir('recieved_files')
+else:
+    for file in os.listdir('recieved_files'):
+        os.remove(os.path.join('recieved_files', file))
 
 
 @bot.message_handler(commands=['start'])
@@ -32,6 +36,28 @@ def send_welcome(message):
                                       "<a href='https://t.me/R_Galiullin'>Галлиулин Руслан</a>", parse_mode="HTML")
 
 
+@bot.message_handler(content_types=['voice'])
+def voice_message(message):
+    print("dd")
+    file_info = bot.get_file(message.voice.file_id)
+    path_to_file = 'recieved_files/' + str(file_info.file_unique_id) + split_string + 'voice.ogg'
+    downloaded_file = bot.download_file(file_info.file_path)
+    with open(path_to_file, 'wb') as f:
+        f.write(downloaded_file)
+
+    AudioSegment.from_ogg(path_to_file).export(path_to_file.split('.')[0] + '.wav', format='wav')
+    os.remove(path_to_file)
+    path_to_file = path_to_file.split('.')[0] + '.wav'
+    extension = 'v'
+
+    files.append(path_to_file)
+    beep = types.InlineKeyboardButton('Запикать', callback_data='b' + extension + str(len(files) - 1))
+    silence = types.InlineKeyboardButton('Приглушить', callback_data='s' + extension + str(len(files) - 1))
+    cancel = types.InlineKeyboardButton('Отменить обработку', callback_data='c' + extension + str(len(files) - 1))
+    buttons = types.InlineKeyboardMarkup([[beep, silence], [cancel]])
+    bot.send_message(message.chat.id, 'Выберите необходимое действие', reply_markup=buttons)
+
+
 @bot.message_handler(content_types=['audio'])
 def message_reply(message):
     file_info = bot.get_file(message.audio.file_id)
@@ -39,14 +65,19 @@ def message_reply(message):
     if not path_to_file.split('.')[1] in ['wav', 'mp3']:
         bot.send_message(message.chat.id, 'Неверный формат файла.')
         return
-
     downloaded_file = bot.download_file(file_info.file_path)
     with open(path_to_file, 'wb') as f:
         f.write(downloaded_file)
+    extension = 'w'
+    if path_to_file.split('.')[1] == 'mp3':
+        AudioSegment.from_mp3(path_to_file).export(path_to_file.split('.')[0] + '.wav', format='wav')
+        os.remove(path_to_file)
+        path_to_file = path_to_file.split('.')[0] + '.wav'
+        extension = 'm'
     files.append(path_to_file)
-    beep = types.InlineKeyboardButton('Запикать', callback_data='b' + str(len(files) - 1))
-    silence = types.InlineKeyboardButton('Приглушить', callback_data='s' + str(len(files) - 1))
-    cancel = types.InlineKeyboardButton('Отменить обработку', callback_data='c' + str(len(files) - 1))
+    beep = types.InlineKeyboardButton('Запикать', callback_data='b' + extension + str(len(files) - 1))
+    silence = types.InlineKeyboardButton('Приглушить', callback_data='s' + extension + str(len(files) - 1))
+    cancel = types.InlineKeyboardButton('Отменить обработку', callback_data='c' + extension + str(len(files) - 1))
     buttons = types.InlineKeyboardMarkup([[beep, silence], [cancel]])
     bot.send_message(message.chat.id, 'Выберите необходимое действие', reply_markup=buttons)
 
@@ -57,24 +88,44 @@ def answer_callback_query(callback_query: types.CallbackQuery):
     bot.delete_message(chat_id=callback_query.message.chat.id,
                        message_id=callback_query.message.message_id)
     bot.send_message(callback_query.from_user.id, 'Выполняется обработка файла')
-    file_path = files[int(callback_query.data[1:])]
-    with open(file_path, 'rb') as file:
-        f = file.read()
+    file_path = files[int(callback_query.data[2:])]
+    if (os.path.isfile(file_path)):
 
-    # обработка файла
+        # обработка файла
 
-    bot.send_document(callback_query.from_user.id, f, visible_file_name=file_path.split(split_string)[1])
-    os.remove(file_path)
+        if callback_query.data[1] == 'm':
+            AudioSegment.from_wav(file_path).export(file_path.split('.')[0] + '.mp3', format='mp3')
+            os.remove(file_path)
+            file_path = file_path.split('.')[0] + '.mp3'
+
+        if callback_query.data[1] == 'v':
+            AudioSegment.from_wav(file_path).export(file_path.split('.')[0] + '.ogg', format='ogg')
+            os.remove(file_path)
+            file_path = file_path.split('.')[0] + '.ogg'
+        print(file_path)
+        file = open(file_path, 'rb')
+
+        if (callback_query.data[1] == 'v'):
+            bot.send_voice(callback_query.from_user.id, file)
+        else:
+            bot.send_document(callback_query.from_user.id, file, visible_file_name=file_path.split(split_string)[1])
+        file.close()
+        os.remove(file_path)
+    else:
+        bot.send_message(callback_query.message.chat.id, 'Файл удалён из системы.')
 
 
 @bot.callback_query_handler(func=lambda c: c.data[0] == "c")
 def answer_callback_query(callback_query: types.CallbackQuery):
     bot.answer_callback_query(callback_query.id)
-    file_path = files[int(callback_query.data[1:])]
-    bot.delete_message(chat_id=callback_query.message.chat.id,
-                       message_id=callback_query.message.message_id)
-    bot.send_message(callback_query.from_user.id, 'Обработка отменена')
-    os.remove(file_path)
+    file_path = files[int(callback_query.data[2:])]
+    if os.path.isfile(file_path):
+        bot.delete_message(chat_id=callback_query.message.chat.id,
+                           message_id=callback_query.message.message_id)
+        bot.send_message(callback_query.from_user.id, 'Обработка отменена')
+        os.remove(file_path)
+    else:
+        bot.send_message(callback_query.message.chat.id, 'Файл удалён из системы.')
 
 
 bot.infinity_polling(none_stop=True)
